This is interface bug fix required. The effect is little. [Before] void* (*alloc_memory_block_zero)(SPVM_ENV* env, int64_t byte_size);
[After]
void* (*alloc_memory_block_zero)(SPVM_ENV* env, size_t byte_size);

+++ lib/SPVM/Builder/include/spvm_api.h
@@ -186,7 +186,7 @@ void SPVM_API_leave_scope(SPVM_ENV* env, int32_t original_mortal_stack_top);
 int32_t SPVM_API_call_spvm_method(SPVM_ENV* env, int32_t method_id, SPVM_VALUE* stack);
 int32_t SPVM_API_call_spvm_method_vm(SPVM_ENV* env, int32_t method_id, SPVM_VALUE* stack);
 
-void* SPVM_API_alloc_memory_block_zero(SPVM_ENV* env, int64_t byte_size);
+void* SPVM_API_alloc_memory_block_zero(SPVM_ENV* env, size_t byte_size);
 void SPVM_API_free_memory_block(SPVM_ENV* env, void* block);
 void* SPVM_API_safe_malloc_zero(int64_t byte_size);
